Lazy-load embedded vid component
